+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ### Changes
 * Boolean literals are now keywords; this means that `true` and `false` are now `@true` and `@false`. The original reserved fragments are now treated like strings.
   * In addition, boolean keywords now ignore surrounding whitespace.
+* Optional parameters without default arguments now no longer define variables in the function body when the user doesn't pass a value to them; as a result, access to these arguments is now fallible.
 
 ### Fixes
 * Fixed panic that occurred with some variadic function calls